requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.11
djangorestframework==3.14.0
idna==3.6
lxml==5.1.0
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
validators==0.22.0
